backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# External API clients
deepgram-sdk>=3.0.0
google-generativeai==0.3.2
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

